Different repr for possible tetrimino states
Previously each state was a string of 4 lines and we converted to
blocks by iterating over lines and checking chars each time. Basically
we were just carrying the format of our data at rest over to the
application itself when it wasn't really that convenient of a
representation. Now instead of a String we need to split into lines
each time, `states!` returns `Vec<Vec<bool>>` to represent the
possible rotation states.