Switch x and y fields on Tetrimino to i32
These numbers represent the offset at which the tetrimino
states we load from the assets/states files should be
drawn (as measured in cells). Because the "well" the
tetriminos fall down is a lot like a 2D array of cells,
I made these numbers `usize`s early on.

Because they represent the offsets of the 4x4 vectors of
bools, it's possible that to get all the blocks in a given
shape to column zero we actually end up with a negative offset
here. For instance, the I shape in vertical alignment:

0100
0100
0100
0100

The Tetrimino x and y values encode the offset of the top left
zero in this shape array. for the 1s to be in the zeroth column
in the well, the offset we store has to be -1 (or we need to
address this is in a different way by always defining shape
offsets in terms of a cell with a 1 that is invariant under
rotation or add hidden buffer columns and do like 1-based
indexing).

IMO, the most straight-forward way to handle this was to just
allow negative offsets while maintaining the restriction that
all Block x and y fields themselves must be >= 0.